# src/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.CreateOrganizationRequest:
    properties:
      description:
        maxLength: 1000
        type: string
      name:
        maxLength: 255
        type: string
      organization_code:
        description: 비어있으면 자동 생성
        type: string
      parent_id:
        description: nil = 최상위 조직
        type: integer
    required:
    - name
    type: object
...
  model.Organization:
    properties:
      children:
        items:
          $ref: '#/definitions/model.Organization'
        type: array
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      organization_code:
        type: string
      parent:
        allOf:
        - $ref: '#/definitions/model.Organization'
        description: 관계 정의 (API 응답 전용 - 필요 시 Preload)
      parent_id:
        type: integer
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/model.User'
        type: array
    type: object
  model.OrganizationDeletableResponse:
    properties:
      deletable:
        type: boolean
      reason:
        type: string
    type: object
  model.OrganizationTree:
    properties:
      children:
        items:
          $ref: '#/definitions/model.OrganizationTree'
        type: array
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      level:
        type: integer
      name:
        type: string
      organization_code:
        type: string
      parent_id:
        type: integer
      path:
        description: '예: "/조직A/개발팀"'
        type: string
      updated_at:
        type: string
      user_count:
        type: integer
    type: object
...
host: localhost
info:
  contact: {}
  description: MC IAM Manager API Documentation
  title: MC IAM Manager API
  version: 0.6.2
paths:
...
  /api/organizations:
    get:
      description: 전체 조직 목록을 조회합니다. tree=true이면 Tree 구조로 반환. name/code로 검색 가능 (검색
        시 tree 파라미터 무시).
      operationId: listOrganizations
      parameters:
      - description: 'Tree 구조 반환 여부 (기본: false)'
        in: query
        name: tree
        type: boolean
      - description: 조직명 검색 (부분 일치, ILIKE)
        in: query
        name: name
        type: string
      - description: 조직 코드 검색 (부분 일치, ILIKE)
        in: query
        name: code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.OrganizationTree'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 조직 목록 조회
      tags:
      - organizations
    post:
      consumes:
      - application/json
      description: 플랫폼 관리자가 조직을 생성합니다. parent_id가 없으면 최상위 조직 생성.
      operationId: createOrganization
      parameters:
      - description: 조직 생성 요청
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CreateOrganizationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Organization'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 조직 생성
      tags:
      - organizations
  /api/organizations/code/{code}:
    get:
      description: 조직 코드로 조직 정보를 조회합니다.
      operationId: getOrganizationByCode
      parameters:
      - description: '조직 코드 (예: 0101)'
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Organization'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 조직 상세 조회 (코드)
      tags:
      - organizations
  /api/organizations/id/{organizationId}:
    delete:
      description: 조직을 삭제합니다. cascade=true이면 하위 조직 및 사용자 매핑도 모두 삭제됩니다. 기본(cascade=false)이고
        하위 조직이 있으면 400을 반환합니다.
      operationId: deleteOrganization
      parameters:
      - description: 조직 ID
        in: path
        name: organizationId
        required: true
        type: integer
      - description: '하위 조직 cascade 삭제 여부 (기본: false)'
        in: query
        name: cascade
        type: boolean
      produces:
      - application/json
      responses:
        "204":
          description: 삭제 성공
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 조직 삭제
      tags:
      - organizations
    get:
      description: 조직 ID로 조직 정보를 조회합니다.
      operationId: getOrganizationByID
      parameters:
      - description: 조직 ID
        in: path
        name: organizationId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Organization'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 조직 상세 조회 (ID)
      tags:
      - organizations
    put:
      consumes:
      - application/json
      description: 조직 정보를 수정합니다. 부모 변경 시 하위 조직 코드 자동 재생성.
      operationId: updateOrganization
      parameters:
      - description: 조직 ID
        in: path
        name: organizationId
        required: true
        type: integer
      - description: 조직 수정 요청
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateOrganizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 조직 수정
      tags:
      - organizations
  /api/organizations/id/{organizationId}/deletable:
    get:
      description: 특정 조직의 삭제 가능 여부와 사유를 반환합니다.
      operationId: getOrganizationDeletable
      parameters:
      - description: 조직 ID
        in: path
        name: organizationId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.OrganizationDeletableResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 조직 삭제 가능 여부 확인
      tags:
      - organizations
  /api/organizations/id/{organizationId}/move:
    put:
      consumes:
      - application/json
      description: 조직을 트리 내 다른 위치(다른 부모 조직)로 이동합니다. 하위 조직도 함께 이동하며 조직 코드가 자동 재생성됩니다.
      operationId: moveOrganization
      parameters:
      - description: 조직 ID
        in: path
        name: organizationId
        required: true
        type: integer
      - description: '이동 요청 (new_parent_id: null이면 최상위로 이동)'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.MoveOrganizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 조직 이동
      tags:
      - organizations
  /api/organizations/id/{organizationId}/subtree:
    get:
      description: 지정 조직을 루트로 하는 하위 트리를 반환합니다.
      operationId: getOrganizationSubtree
      parameters:
      - description: 조직 ID
        in: path
        name: organizationId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.OrganizationTree'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 특정 조직의 하위 트리 조회
      tags:
      - organizations
  /api/organizations/id/{organizationId}/users:
    get:
      description: 특정 조직에 소속된 사용자 목록을 조회합니다.
      operationId: getOrganizationUsers
      parameters:
      - description: 조직 ID
        in: path
        name: organizationId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 조직 소속 사용자 조회
      tags:
      - organizations
  /api/organizations/tree:
    get:
      description: 전체 조직을 계층 트리 구조로 반환합니다. 최대 10단계 깊이를 지원하며 각 노드에 children 배열을 포함합니다.
      operationId: getOrganizationTree
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.OrganizationTree'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 전체 조직 트리 조회
      tags:
      - organizations
...
  /api/setup/initial-organizations:
    post:
      description: YAML 시드 파일에서 기본 조직 구조(MZC + 8개 프레임워크)를 로드하여 등록합니다. 멱등성 보장.
      operationId: setupInitialOrganizations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 기본 조직 초기화
      tags:
      - organizations
...
  /api/setup/migrate-kc-group-identifiers:
    post:
      description: 각 조직의 Keycloak 그룹 식별자를 레거시 조직명 기준에서 유일성이 보장되는 organization_code
        기준으로 이관합니다(IAM-BUG-029). 멱등성 보장 — 여러 번 호출해도 안전합니다.
      operationId: migrateKeycloakGroupIdentifiers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "207":
          description: 일부 조직 마이그레이션 실패 — errors 필드 참고
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Keycloak 그룹 식별자 마이그레이션
      tags:
      - organizations
...
  /api/users/id/{userId}/organizations/{organizationId}:
    delete:
      description: 사용자를 특정 조직에서 제거합니다.
      operationId: removeUserOrganization
      parameters:
      - description: 사용자 ID
        in: path
        name: userId
        required: true
        type: integer
      - description: 조직 ID
        in: path
        name: organizationId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 사용자-조직 매핑 제거
      tags:
      - organizations
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"